# Speed-up of model eval