--- a/src/common/evaluation/QA/environment.yml
+++ b/src/common/evaluation/QA/environment.yml
@@ -2,12 +2,12 @@ name: CGM_QA_Pipeline
 channels:
   - conda-forge
 dependencies:
-  - python=3.6
+  - python=3.7
   - nb_conda_kernels==2.2.1
   - scikit-image
   - pip:
     - azure-cli==2.0.69
-    - azureml-sdk[notebooks,automl]==1.13.0
+    - azureml-sdk[notebooks,automl]==1.18.0
     - papermill==1.0.1
     - nbconvert==5.6.0
     - matplotlib==3.2.1