--- a/src/common/evaluation/QA/test-pipeline.yml
+++ b/src/common/evaluation/QA/test-pipeline.yml
@@ -3,8 +3,6 @@ trigger:
     include:
     - releases/*
 
-pr: none
-
 jobs:
 - job: EvaluationJob
   timeoutInMinutes: 300