**Findings:** Mounting the data slows down prediction time because the data has to be fetched over the network. It is a lot faster to fully download the data before starting to predict.

**Changes:**
- Download instead of mounting (to speed up)
- Use batch size of 1024 instead of 256 (to speed up)
- remove caching since we use every sample only once in model evaluation (unlike in training)
- add some print statements to mention how fast the prediction
- Use TF2.3 and py3.7

**Details about speed measurements**

Before this PR: The run takes 32m 10.37s ([Run 204](https://ml.azure.com/experiments/id/4347c355-aa9d-4482-b7d9-2130b9aa7c44/runs/QA-pipeline_1606901417_9803b4cd?wsid=/subscriptions/9b82ecea-6780-4b85-8acf-d27d79028f07/resourceGroups/cgm-ml-prod/providers/Microsoft.MachineLearningServices/workspaces/cgm-azureml-prod&tid=006dabd7-456d-465b-a87f-f7d557e319c8),  [pipeline](https://dev.azure.com/cgmorg/ChildGrowthMonitor/_build/results?buildId=1658&view=logs&j=3d42b299-58dd-525f-2f90-47c9dbee9ebd&t=66a92730-1935-5d8a-99af-e052f481c0da))

We tried downloading data instead of mounting:
- Download data first: 8m 58.57s [Run 254](https://ml.azure.com/experiments/id/4347c355-aa9d-4482-b7d9-2130b9aa7c44/runs/QA-pipeline_1606990079_adfd3b4a?wsid=/subscriptions/9b82ecea-6780-4b85-8acf-d27d79028f07/resourceGroups/cgm-ml-prod/providers/Microsoft.MachineLearningServices/workspaces/cgm-azureml-prod&tid=006dabd7-456d-465b-a87f-f7d557e319c8), [pipeline](https://dev.azure.com/cgmorg/ChildGrowthMonitor/_build/results?buildId=1711&view=logs&j=3d42b299-58dd-525f-2f90-47c9dbee9ebd&t=66a92730-1935-5d8a-99af-e052f481c0da&l=14): Total time for prediction experiment: `206.96 sec = 3.45min`
-> This is a great improvement

Then we we tried to disable the GPU:
- Download data first, disable GPU: 18m 6.555s [Run 261](https://ml.azure.com/experiments/id/4347c355-aa9d-4482-b7d9-2130b9aa7c44/runs/QA-pipeline_1606992601_6e442285?wsid=/subscriptions/9b82ecea-6780-4b85-8acf-d27d79028f07/resourceGroups/cgm-ml-prod/providers/Microsoft.MachineLearningServices/workspaces/cgm-azureml-prod&tid=006dabd7-456d-465b-a87f-f7d557e319c8#outputsAndLogs): Total time for prediction experiment: `751.55 sec = 12.5min` 
-> This is actually worse so GPU seems to be enabled correctly (This is also what the logs show)

We tried to increase the batch size: 
- TF2.3, mini, predict from tf.data (debug=True, bs=512): 34s
- TF2.3, mini, predict from tf.data (debug=True, bs=1024): OOM

Before merging: Revert `pr: none`